Use gin to handle routing, make getBooks a get request that accepted a title query param. The title query param is then used to scrape books from goodreads of that title
